refactor(cli): retire the legacy folder
src/legacy/ held the last types named after the deleted command shell. CliRuntime and CommandContext move to controllers/context.ts, CommandSuccess moves into its only user, controllers/app-env-file.ts, and the unused CliOutput is gone. The legacyOperationContext adapter is now operationContext, and its test file follows the rename.
